fix/productImage : 순환참조 문제 해결을 위해 클래스 분리 (Split classes to eliminate circular dependency)
