+++ src/main/java/org/example/lastcall/domain/product/controller/ProductImageController.java
@@ -6,6 +6,7 @@ import org.example.lastcall.domain.product.dto.request.ProductImageCreateRequest
 import org.example.lastcall.domain.product.dto.response.ProductImageResponse;
 import org.example.lastcall.domain.product.sevice.ProductCommandServiceApi;
 import org.example.lastcall.domain.product.sevice.ProductImageService;
+import org.example.lastcall.domain.product.sevice.ProductImageViewService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +18,8 @@ import java.util.List;
 @RequestMapping("/api/v1/products")
 public class ProductImageController {
     private final ProductImageService productImageService;
-    private final ProductCommandServiceApi productCommandServiceApi;
+    private final ProductImageViewService productImageViewService;
+    private final ProductCommandServiceApi productCommandServiceApi; //순환 참조 막기 위해 추가
 
     @PostMapping("/{productId}/images")
     public ResponseEntity<ApiResponse<List<ProductImageResponse>>> createProductImage(@PathVariable Long productId,
@@ -42,7 +44,7 @@ public class ProductImageController {
     //상품별 이미지 전체 조회
     @GetMapping("/{productId}/image")
     public ResponseEntity<ApiResponse<List<ProductImageResponse>>> readAllProductImage(@PathVariable Long productId) {
-        List<ProductImageResponse> response = productImageService.readAllProductImage(productId);
+        List<ProductImageResponse> response = productImageViewService.readAllProductImage(productId);
         ApiResponse<List<ProductImageResponse>> apiResponse = ApiResponse.success("상품별 이미지 조회에 성공했습니다.", response);
 
         return ResponseEntity.ok(apiResponse);
