+++ src/main/java/org/example/lastcall/domain/product/sevice/ProductViewServiceApi.java
@@ -3,7 +3,7 @@ package org.example.lastcall.domain.product.sevice;
 import org.example.lastcall.domain.product.dto.response.ProductResponse;
 
 // 반드시 ProductViewServiceApi로 이름 바꿀 것!
-public interface ProductServiceApi {
+public interface ProductViewServiceApi {
     //상품 단건 조회
     ProductResponse readProduct(Long productId);
